requirements.txt
cdp-langchain
python-dotenv
ethers
pandas
langchain-experimental
tabulate
chromadb
flask
requests
flask-cors
celery
redis